requirements.txt
opencv-python
ultralytics
streamlit
google-cloud-vision
numpy
